requirements.txt
streamlit==1.30.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.1.1
numpy==1.26.0
datasets==2.16.1
scikit-learn==1.3.2
transformers==4.33.3
huggingface-hub==0.20.0